surya/requirements.txt
surya-ocr
Pillow
